+++ CHANGELOG.md
@@ -7,3 +7,14 @@ and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- [DAGPLEJ-62](https://jira.itkdev.dk/browse/DAGPLEJ-62)
+  Added day carer name and address lookup.
+- [DAGPLEJ-20](https://jira.itkdev.dk/browse/DAGPLEJ-20):
+  Added OpenID Connect module and configuration.
+- [DAGPLEJ-21](https://jira.itkdev.dk/browse/DAGPLEJ-21):
+  Added Drupal Core, base config and Drupal Commerce.
+- [DAGPLEJ-30](https://jira.itkdev.dk/browse/DAGPLEJ-30):
+  - Added eStore theme, and default product entity.
+  - Added front page view
+  - Added product category